backend/requirements.txt
Django==3.2.8
djangorestframework==3.14.0
Pillow==9.2.0
whitenoise==6.2.0
django-cors-headers
djangorestframework-simplejwt
psycopg2
psycopg2-binary
